pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pc-rai"
version = "0.1.0"
description = "Point Cloud Rockfall Activity Index Classification Tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Connor", email = "connor@example.com"}
]
keywords = [
    "lidar",
    "point-cloud",
    "geomorphology",
    "rockfall",
    "hazard-assessment",
    "coastal-erosion",
    "classification",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "numpy>=1.21",
    "scipy>=1.7",
    "laspy[lazrs]>=2.4",
    "matplotlib>=3.5",
    "tqdm>=4.62",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "build>=0.10",
]
viz = [
    # Open3D for 3D visualization (requires Python <3.13)
    "open3d>=0.17; python_version<'3.13'",
]

[project.scripts]
pc-rai = "pc_rai.cli:main"

[project.urls]
Homepage = "https://github.com/username/pc-rai"
Documentation = "https://github.com/username/pc-rai#readme"
Repository = "https://github.com/username/pc-rai.git"
Issues = "https://github.com/username/pc-rai/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["pc_rai*"]

[tool.setuptools.package-data]
pc_rai = ["py.typed"]

# =============================================================================
# Black Configuration
# =============================================================================
[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# =============================================================================
# Ruff Configuration
# =============================================================================
[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "ARG001", # unused function argument (common in callbacks)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "S101"]  # Allow assert and unused args in tests

[tool.ruff.lint.isort]
known-first-party = ["pc_rai"]

# =============================================================================
# Mypy Configuration
# =============================================================================
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Gradually enable
ignore_missing_imports = true
exclude = [
    "tests/",
    "build/",
    "dist/",
]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

# =============================================================================
# Pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "requires_cloudcompare: marks tests that require CloudCompare installation",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# =============================================================================
# Coverage Configuration
# =============================================================================
[tool.coverage.run]
source = ["pc_rai"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 70
show_missing = true
